pyproject.toml
[tool.poetry]
name = "login-api"
version = "0.1.0"
description = ""
authors = ["Pax Prz <paxprz@whoami.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-RESTful = "^0.3.8"
SQLAlchemy = "^1.3.22"
Flask-Script = "^2.0.6"
python-dotenv = "^0.15.0"
python-jose = "^3.2.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
